refactor(backend): extract ConversationQueries from OpenBotDatabase
Moves every read of a thread's messages — the whole snapshot, an anchored page,
a page's supported cursor and thread search — into
`src/backend/database/conversation-queries.ts`, together with the opaque
base64url cursor helpers and `escapeLike`.

`conversationMarkerSqlFilter` moves with the three
`*_EVENT_ITEM_TYPE_PREFIX` constants it interpolates rather than being left
behind at its callers: a filter interpolating `undefined` still runs and quietly
returns the wrong rows.

Every default argument is duplicated on the facade delegation — `anchor`,
`requestedLimit`, `options` — so a caller that omits one still reaches the
controller with a value.
